stages/keymap: preserve values not set by systemd-firstboot in vconsole.conf
The stage currently removes the `/etc/vconsole.conf` before running
`systemd-firstboot`. The result is that some original values from the
`vconsole.conf` file are lost, e.g. the 'FONT' option.

Modify the stage to first read the key-values from `vconsole.conf` if it
exists. Then run `systemd-firstboot` with the `--force` option, which
ensures that the file is replaced even if it already exists. Then read
the content of the configuration file again and update values originally
read from the configuration files with the ones set by
`systemd-firstboot`. As the last step write the merged key-values to
`vconsole.conf` file.

Signed-off-by: Tomas Hozza <thozza@redhat.com>