osbuild/solver/dnf5.py: include empty "modules" object for depsolve
Make the DNF4 and DNF5 depsolve command response consistent by always
returning an empty "modules" object in the DNF5 implementation. This is
compliant with the DNF4 implementation in case no modules are in the
depsolved package set. Since DNF5 implementation does not support
depsolving with module-specs, the returned "modules" object will be
always empty.

Note that this does not affect the solver API, since DNF4 implementation
is already returning the "modules" object and the osbuild/images
implementation always expects the "modules" to be set in the response.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>