tools/test/depsolve: merge depsolve() API test into test_depsolve()
Transform test_depsolve_result_api() test case into a helper function
which will be used to check the API response in other test cases,
specifically in the test_depsolve() test case to begin with. The reason
for this is to de-duplicate the same checks and to make it easier in the
future to change the API and keep the testing sane. Otherwise, we would
have to implement conditional check in every place that inspects the
depsolve API response.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>